+++ backend/db/models/spot.js
@@ -50,7 +50,7 @@ module.exports = (sequelize, DataTypes) => {
 				allowNull: false,
 			},
 			lat: {
-				type: DataTypes.DECIMAL,
+				type: DataTypes.FLOAT,
 				allowNull: false,
 				validate: {
 					min: -90,
@@ -58,7 +58,7 @@ module.exports = (sequelize, DataTypes) => {
 				},
 			},
 			lng: {
-				type: DataTypes.DECIMAL,
+				type: DataTypes.FLOAT,
 				allowNull: false,
 				validate: {
 					min: -180,
@@ -74,7 +74,7 @@ module.exports = (sequelize, DataTypes) => {
 				allowNull: false,
 			},
 			price: {
-				type: DataTypes.DECIMAL,
+				type: DataTypes.FLOAT,
 				allowNull: false,
 				validate: {
 					min: 0,
